:art: Refactor: 지도 추천장소 함수 리팩토링
